fix: resolve all PostgreSQL test failures for full DB compatibility
- Use valid UUID format in ApiAuthenticationTest (PostgreSQL rejects
  non-UUID strings for uuid columns)
- Add code filter to LocaleDataSource and update filter test to use
  code instead of boolean status (PostgreSQL is strict on boolean types)
- Generate option code in AttributeRepository when creating options
  without code (PostgreSQL enforces NOT NULL strictly)
- Cast boolean status to int in ImportProductsAclToolTest (PostgreSQL
  returns true/false, MySQL returns 1/0)

Both MySQL and PostgreSQL: 1169 passed, 0 failed.

+++ packages/Webkul/Admin/tests/Feature/MagicAi/ImportProductsAclToolTest.php
@@ -91,7 +91,7 @@ it('skips existing products during mixed AI import when the user lacks edit perm
         ->and(data_get($result, 'result.skipped'))->toBe(1)
         ->and(data_get($result, 'result.errors.0'))->toContain('catalog.products.edit');
 
-    expect(Product::query()->where('sku', 'AI-EXISTING-100')->value('status'))->toBe(0)
+    expect((int) Product::query()->where('sku', 'AI-EXISTING-100')->value('status'))->toBe(0)
         ->and(Product::query()->where('sku', 'AI-NEW-101')->exists())->toBeTrue();
 });
